Move jar file version extractor to separate file
